requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
openai-whisper @ git+https://github.com/openai/whisper.git
PyAudio==0.2.14
keyboard==0.13.5
requests==2.32.5
